+++ config/scramble-extensions.php
@@ -6,6 +6,7 @@ return [
         'trait' => 'Mitoop\\Http\\RespondsWithJson',
         'model_namespace' => 'App\\Models',
         'column_comments' => true,
+        'schema_overrides' => [],
     ],
 
     'scene_form_request' => [